adds calling back the callback function when the transfer is done.
The RGBBus doesn't have the ability to set a callback function when transferring data to the display. I am not sure why it doesn't so if there is a callback that has been supplied that callback gets called after the function to transfer the data returns. We will see if that works out properly