updates make.py
I added a new parameter for the esp32's it is "-skip-partition-resize". This is put in there because now a resize of the partitions is going to occur every time the build takes place. The partition size gets sized up and also sized down. So if you have not made any changes that would effect the size of the firmware then you can specify this flag to shorten up the build process a little.

After resizing the partition the build no longer has to run the full amount. It is much shorter than needing to compile the entire project again.

I also have the script automatically disable I2S. This is due to an upstream bug. I reported the bug so hopefully in the next release of MicroPython it will be fixed. There is also a big in btree that is upstream as well. This has also been reported and again I hope it is fixed in the next version.

I added some output text that appears in red to let the user know what the build is doing. It makes it easier to see what is happening.

I added special handling of the deploy build argument. This is only going to be appended to the build command after the partition size has been updated OR if the resizing is skipped. And it will only be appended if the user has added it when running make.py